+++ index.js
@@ -1,30 +1,5 @@
 #!/usr/bin/env node
 // transpile:main
-
-import { asyncify } from 'asyncbox';
-import * as server from './lib/server';
-
-
-const { startServer } = server;
-
-const DEFAULT_HOST = 'localhost';
-const DEFAULT_PORT = 4723;
-
-async function main () {
-  const getArgValue = (argName) => {
-    const argIndex = process.argv.indexOf(argName);
-    return argIndex > 0 ? process.argv[argIndex + 1] : null;
-  };
-  const port = parseInt(getArgValue('--port'), 10) || DEFAULT_PORT;
-  const host = getArgValue('--host') || DEFAULT_HOST;
-  return await startServer(port, host);
-}
-
-if (require.main === module) {
-  asyncify(main);
-}
-
-
 import * as driver from './lib/driver';
 import * as androidHelperIndex from './lib/android-helpers';
 import * as commandIndex from './lib/commands/index';
