+++ lib/android-helpers.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import path from 'path';
 import { retry, waitForCondition } from 'asyncbox';
 import logger from './logger';
-import { fs, util } from '@appium/support';
+import { fs, util } from 'appium/support';
 import { path as settingsApkPath } from 'io.appium.settings';
 import Bootstrap from './bootstrap';
 import B from 'bluebird';
