+++ lib/bootstrap.js
@@ -2,9 +2,9 @@ import UiAutomator from './uiautomator';
 import net from 'net';
 import path from 'path';
 import _ from 'lodash';
-import { errorFromCode } from '@appium/base-driver';
+import { errorFromCode } from 'appium/driver';
 import B from 'bluebird';
-import { logger } from '@appium/support';
+import { logger } from 'appium/support';
 
 
 const log = logger.getLogger('AndroidBootstrap');
