+++ lib/commands/actions.js
@@ -1,5 +1,5 @@
 import androidHelpers from '../android-helpers';
-import { fs, util, tempDir} from '@appium/support';
+import { fs, util, tempDir} from 'appium/support';
 import path from 'path';
 import B from 'bluebird';
 import jimp from 'jimp';
