+++ lib/commands/alert.js
@@ -1,5 +1,5 @@
 /* eslint-disable require-await */
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 let commands = {}, helpers = {}, extensions = {};
