+++ lib/commands/app-management.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 import { waitForCondition } from 'asyncbox';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 import { APP_STATE } from '../android-helpers';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 const APP_EXTENSIONS = ['.apk', '.apks'];
