+++ lib/commands/context.js
@@ -2,8 +2,8 @@ import _ from 'lodash';
 import Chromedriver from 'appium-chromedriver';
 import PortFinder from 'portfinder';
 import B from 'bluebird';
-import { util } from '@appium/support';
-import { errors } from '@appium/base-driver';
+import { util } from 'appium/support';
+import { errors } from 'appium/driver';
 import {
   default as webviewHelpers,
   NATIVE_WIN, WEBVIEW_BASE, WEBVIEW_WIN, CHROMIUM_WIN, KNOWN_CHROME_PACKAGE_NAMES
