+++ lib/commands/element.js
@@ -1,6 +1,6 @@
 import androidHelpers from '../android-helpers';
 import { retryInterval } from 'asyncbox';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 
 
 let commands = {}, helpers = {}, extensions = {};
