+++ lib/commands/emu-console.js
@@ -1,4 +1,4 @@
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 const EMU_CONSOLE_FEATURE = 'emulator_console';
