+++ lib/commands/execute.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { errors, PROTOCOLS } from '@appium/base-driver';
+import { errors, PROTOCOLS } from 'appium/driver';
 
 const extensions = {};
