+++ lib/commands/file-actions.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
-import { fs, util, zip, tempDir } from '@appium/support';
+import { fs, util, zip, tempDir } from 'appium/support';
 import path from 'path';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 
 const CONTAINER_PATH_MARKER = '@';
