+++ lib/commands/find.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { errors, isErrorType } from '@appium/base-driver';
+import { errors, isErrorType } from 'appium/driver';
 
 
 let helpers = {}, extensions = {};
@@ -44,7 +44,7 @@ helpers.findElOrEls = async function findElOrEls (strategy, selector, mult, cont
 
       // now we have to inspect the error to determine if it is a no such
       // element error, based on the shape of the error object from
-      // @appium/base-driver
+      // appium/driver
       if (isErrorType(err, errors.NoSuchElementError)) {
         // we are fine with this, just indicate a retry
         return false;
