+++ lib/commands/general.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 import androidHelpers from '../android-helpers';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 import moment from 'moment';
 import { longSleep } from 'asyncbox';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 const MOMENT_FORMAT_ISO8601 = 'YYYY-MM-DDTHH:mm:ssZ';
