+++ lib/commands/intent.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 const NO_VALUE_ARG_TYPE = 'sn';
 const SUPPORTED_EXTRA_TYPES = [
