+++ lib/commands/log.js
@@ -2,7 +2,7 @@ import log from '../logger';
 import os from 'os';
 import _ from 'lodash';
 import WebSocket from 'ws';
-import { DEFAULT_WS_PATHNAME_PREFIX, BaseDriver } from '@appium/base-driver';
+import { DEFAULT_WS_PATHNAME_PREFIX, BaseDriver } from 'appium/driver';
 
 const GET_SERVER_LOGS_FEATURE = 'get_server_logs';
