+++ lib/commands/network.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 import B from 'bluebird';
 
 let commands = {}, helpers = {}, extensions = {};
