+++ lib/commands/recordscreen.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { waitForCondition } from 'asyncbox';
-import { util, fs, net, tempDir, system, timing } from '@appium/support';
+import { util, fs, net, tempDir, system, timing } from 'appium/support';
 import { exec } from 'teen_process';
 import path from 'path';
