+++ lib/commands/shell.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { exec } from 'teen_process';
-import { util } from '@appium/support';
-import { errors } from '@appium/base-driver';
+import { util } from 'appium/support';
+import { errors } from 'appium/driver';
 
 const ADB_SHELL_FEATURE = 'adb_shell';
