+++ lib/commands/streamscreen.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { fs, system, logger, util } from '@appium/support';
+import { fs, system, logger, util } from 'appium/support';
 import { exec, SubProcess } from 'teen_process';
 import { checkPortStatus } from 'portscanner';
 import http from 'http';
