+++ lib/commands/touch.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 import androidHelpers from '../android-helpers';
 import B from 'bluebird';
-import { errors, isErrorType } from '@appium/base-driver';
+import { errors, isErrorType } from 'appium/driver';
 import { asyncmap } from 'asyncbox';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 
 let commands = {}, helpers = {}, extensions = {};
