+++ lib/driver.js
@@ -1,4 +1,4 @@
-import { BaseDriver, DeviceSettings } from '@appium/base-driver';
+import { BaseDriver, DeviceSettings } from 'appium/driver';
 import desiredConstraints from './desired-caps';
 import commands from './commands/index';
 import {
@@ -7,7 +7,7 @@ import {
 } from './android-helpers';
 import _ from 'lodash';
 import { DEFAULT_ADB_PORT } from 'appium-adb';
-import { fs, tempDir, util } from '@appium/support';
+import { fs, tempDir, util } from 'appium/support';
 import { retryInterval } from 'asyncbox';
 import { SharedPrefsBuilder } from 'shared-preferences-builder';
 import B from 'bluebird';
