+++ lib/uiautomator.js
@@ -1,5 +1,5 @@
 import events from 'events';
-import { logger } from '@appium/support';
+import { logger } from 'appium/support';
 
 
 const log = logger.getLogger('UiAutomator');
