+++ lib/unlock-helpers.js
@@ -1,7 +1,7 @@
 import logger from './logger';
 import { sleep } from 'asyncbox';
 import _ from 'lodash';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 
 const PIN_UNLOCK = 'pin';
 const PIN_UNLOCK_KEY_EVENT = 'pinWithKeyEvent';
