+++ lib/webview-helpers.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import logger from './logger';
 import axios from 'axios';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 import { findAPortNotInUse } from 'portscanner';
 import LRU from 'lru-cache';
 import B from 'bluebird';
