+++ test/functional/commands/find/from-el-e2e-specs.js
@@ -2,7 +2,7 @@ import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import AndroidDriver from '../../../../lib/driver';
 import { DEFAULT_CAPS } from '../../capabilities';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 
 
 chai.should();
