+++ test/functional/webview-browser-tester-e2e-specs.js
@@ -1,11 +1,11 @@
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import { AndroidDriver } from '../../lib/driver';
-import { startServer } from '../../lib/server';
+import { startServer } from '../server';
 import { ensureAVDExists, getChromedriver220Asset } from './helpers';
 import { CHROME_CAPS } from './capabilities';
 import _ from 'lodash';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 
 
 chai.should();
