+++ test/unit/android-helper-specs.js
@@ -3,8 +3,8 @@ import chaiAsPromised from 'chai-as-promised';
 import sinon from 'sinon';
 import helpers, { prepareAvdArgs, ensureNetworkSpeed } from '../../lib/android-helpers';
 import ADB from 'appium-adb';
-import { withMocks } from '@appium/test-support';
-import { fs } from '@appium/support';
+import { withMocks } from 'appium/test-support';
+import { fs } from 'appium/support';
 import unlocker from '../../lib/unlock-helpers';
 import _ from 'lodash';
 import B from 'bluebird';
