+++ test/unit/commands/general-specs.js
@@ -3,8 +3,8 @@ import chaiAsPromised from 'chai-as-promised';
 import sinon from 'sinon';
 import AndroidDriver from '../../../lib/driver';
 import helpers from '../../../lib/android-helpers';
-import { withMocks } from '@appium/test-support';
-import { fs } from '@appium/support';
+import { withMocks } from 'appium/test-support';
+import { fs } from 'appium/support';
 import Bootstrap from '../../../lib/bootstrap';
 import B from 'bluebird';
 import ADB from 'appium-adb';
