+++ test/unit/commands/ime-specs.js
@@ -3,7 +3,7 @@ import chaiAsPromised from 'chai-as-promised';
 import sinon from 'sinon';
 import AndroidDriver from '../../../lib/driver';
 import ADB from 'appium-adb';
-import { errors } from '@appium/base-driver';
+import { errors } from 'appium/driver';
 
 chai.should();
 chai.use(chaiAsPromised);
