+++ test/unit/commands/recordscreen-specs.js
@@ -1,8 +1,8 @@
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import AndroidDriver from '../../../lib/driver';
-import { withMocks } from '@appium/test-support';
-import { fs, tempDir } from '@appium/support';
+import { withMocks } from 'appium/test-support';
+import { fs, tempDir } from 'appium/support';
 import ADB from 'appium-adb';
