+++ test/unit/driver-specs.js
@@ -3,11 +3,11 @@ import chaiAsPromised from 'chai-as-promised';
 import log from '../../lib/logger';
 import sinon from 'sinon';
 import { helpers, SETTINGS_HELPER_PKG_ID } from '../../lib/android-helpers';
-import { withMocks } from '@appium/test-support';
+import { withMocks } from 'appium/test-support';
 import AndroidDriver from '../../lib/driver';
 import ADB from 'appium-adb';
-import { errors } from '@appium/base-driver';
-import { fs } from '@appium/support';
+import { errors } from 'appium/driver';
+import { fs } from 'appium/support';
 import { SharedPrefsBuilder } from 'shared-preferences-builder';
 import _ from 'lodash';
